phase-a: widen H3Index to __uint128_t + Group C macros + cascade fixes
Foundational typedef widening per playbook §3 and POC_TRANSFER_GUIDE.md
Phase A. Stock byte-identity preserved (high half = 0 for stock cells).

Type widening:
- h3api.h.in: typedef uint64_t H3Index → typedef __uint128_t H3Index,
  guarded by #error on missing __SIZEOF_INT128__ (per playbook §9.1).

Group C macros (h3Index.h, copied verbatim from poc1_bit_layout.c):
- H3_EXT_FLAG_OFFSET=64, H3_EXT_FLAG_MASK, H3_EXT_DIGITS_OFFSET=65,
  H3_EXT_DIGITS_MASK, H3_INIT_EXT, H3_GET/SET_EXT_FLAG,
  H3_GET/SET_EFFECTIVE_RESOLUTION, H3_GET/SET_EXT_INDEX_DIGIT,
  H3_GET/SET_DIGIT_AT_RES.
- _Static_assert(sizeof(H3Index) == 16, ...) for gate A1.

Stock Group A mask retype:
- All bit-position masks (HIGH_BIT, MODE, BC, RES, RESERVED, DIGIT) and
  H3_INIT changed from (uint64_t) cast to (H3Index) cast. Bit values are
  unchanged. The retype is necessary because ~MASK in 64-bit integers
  zero-extends to __uint128_t, clobbering the high half on every SET
  macro AND. With (H3Index) typing the mask, ~MASK is full-width and
  preserves bits 64-127.

Cascade fixes (typedef widening cascade through every file using PRIx64):
- src/h3lib/lib/h3Index.c stringToH3: scan into uint64 intermediate, cast
  to H3Index (preserves stock 16-char hex round-trip; Phase C will widen
  to 32-char).
- src/h3lib/lib/h3Index.c h3ToString: cast to (uint64_t) for sprintf
  (stock low-64 hex output; Phase C will widen).
- src/apps/{applib,filters,miscapps,testapps}: (uint64_t) cast on every
  PRIx64/PRIu64 print site touching H3Index. Zero-init for sscanf-target
  H3Index variables in args framework callers (intToString,
  areNeighborCells, cellsToDirectedEdge) so the high half is well-defined
  when sscanf only writes the low 64 bits via PRIx64/PRIu64.

Stock predicate minimal patches (NOT widening — stock-correctness
preservation only; Phase E does the proper ext widening per POC-2):
- _hasAll7AfterRes (playbook §6.4): the bit-magic window
  ~h <<= shift; >>= shift naturally extended from low 30 bits in 64-bit
  to low 94 bits in 128-bit, pulling inverted high-half bits (all 1s for
  stock) into the validation. Restrict to (uint64_t)h to preserve stock
  behavior identically.
- _hasDeletedSubsequence (playbook §6.5): the h <<= 19; >>= 19 window
  naturally extended from low 45 bits to low 109 bits, so _firstOneIndex
  picked up the mode/res metadata in bits 45-63 instead of digit bits.
  Restrict to (uint64_t)h to preserve stock behavior.

Phase A gates:
- A1 sizeof(H3Index)==16: PASS (static_assert in h3Index.h)
- A2 Group C macros compile: PASS (build clean)
- A3 MAX_H3_EXT_RES constant: PASS (already in constants.h)
- A4 316/316 stock ctest: PASS
- A5 ASAN+UBSAN clean on stock suite: PASS (build-dev all green)

+++ src/apps/applib/lib/utility.c
@@ -39,12 +39,14 @@ void error(const char *msg) {
 /**
  * Prints the H3Index
  */
-void h3Print(H3Index h) { printf("%" PRIx64, h); }
+// H3-EXTENDED: print stock low 64 bits only; Phase C will widen to 32-char hex.
+void h3Print(H3Index h) { printf("%" PRIx64, (uint64_t)h); }
 
 /**
  * Prints the H3Index and a newline
  */
-void h3Println(H3Index h) { printf("%" PRIx64 "\n", h); }
+// H3-EXTENDED: print stock low 64 bits only; Phase C will widen to 32-char hex.
+void h3Println(H3Index h) { printf("%" PRIx64 "\n", (uint64_t)h); }
 
 /**
  * Prints the CoordIJK
